Error dato long en chequeo api solucionado
ademas se agregan algunas mejoras esteticas. y se agrega MODO OFFLINE cuando ocurren errores en la api o con archivo config

+++ TatetiC/Registros.c
@@ -279,7 +279,7 @@ int Validar_API(void)
     //Variables para el get
     CURL *curl;
     CURLcode res;
-    int codigo;
+    long codigo;
 
     //Variables para el manejo del archivo
     char linea[MAXLINEA];
